Created code for the logit and Nested Logit (incomplete),
Refactored code so that a number of methods that were previously part of child classes to antitrust are now part of the antitrust class. I also switched around which classes are children of what, giving a more prominent relationship to the Logit Class.

augmented test.R to include tests for pcaids.